# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mesa de Oración API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MensajeRespuesta {
  mensaje: string;
}

export interface Usuario {
  id: number;
  googleId: string;
  nombre: string;
  email: string;
  avatar?: string | null;
  creadoEn: string;
}

export interface UsuarioPublico {
  id: number;
  nombre: string;
  avatar?: string | null;
}

export interface Oracion {
  id: number;
  titulo: string;
  texto: string;
  categoria: string;
  duracionMinutos?: number | null;
  imagen?: string | null;
  autor: UsuarioPublico;
  totalAmenes: number;
  totalMeAyuda: number;
  totalLaRezareHoy: number;
  totalFavoritos: number;
  totalComentarios: number;
  esFavorito: boolean;
  miReaccion?: string | null;
  creadoEn: string;
}

export interface Comentario {
  id: number;
  texto: string;
  autor: UsuarioPublico;
  creadoEn: string;
}

export type OracionDetalle = Oracion & {
  comentarios: Comentario[];
};

export interface ListaOracionesRespuesta {
  oraciones: Oracion[];
  total: number;
  pagina: number;
  totalPaginas: number;
}

export interface CategoriaConteo {
  categoria: string;
  conteo: number;
}

export interface Estadisticas {
  totalOraciones: number;
  totalUsuarios: number;
  totalComentarios: number;
  totalReacciones: number;
}

export interface CrearOracionBody {
  /**
   * @minLength 3
   * @maxLength 200
   */
  titulo: string;
  /** @minLength 10 */
  texto: string;
  categoria: string;
  duracionMinutos?: number | null;
  imagen?: string | null;
}

export interface ActualizarOracionBody {
  /**
   * @minLength 3
   * @maxLength 200
   */
  titulo?: string;
  /** @minLength 10 */
  texto?: string;
  categoria?: string;
  duracionMinutos?: number | null;
  imagen?: string | null;
}

export interface CrearComentarioBody {
  texto: string;
}

export type ReaccionarBodyTipo =
  (typeof ReaccionarBodyTipo)[keyof typeof ReaccionarBodyTipo];

export const ReaccionarBodyTipo = {
  amen: "amen",
  me_ayuda: "me_ayuda",
  la_rezare_hoy: "la_rezare_hoy",
} as const;

export interface ReaccionarBody {
  tipo: ReaccionarBodyTipo;
}

export interface FavoritoRespuesta {
  esFavorito: boolean;
  mensaje: string;
}

export interface ReaccionRespuesta {
  tipo?: string | null;
  mensaje: string;
  totalAmenes: number;
  totalMeAyuda: number;
  totalLaRezareHoy: number;
}

export type CallbackGoogleParams = {
  code?: string;
  error?: string;
};

export type ListarOracionesParams = {
  categoria?: string;
  busqueda?: string;
  pagina?: number;
  limite?: number;
};